updates/debug

+++ quick-spec/spec-m1m2-ta.sh
@@ -30,6 +30,10 @@ spectrumid=${cluster}-${fitid}
 
 ######################################################################
 # convert to ctr = cts/s
+CONVERT_TO_CTR=0
+
+if [[ $CONVERT_TO_CTR -eq 1 ]]
+then
 
 spec=inspec.pha
 
@@ -54,6 +58,8 @@ done
 
 # sleep 200
 
+fi
+
 ######################################################################
 # rebin the spectra
 
@@ -68,14 +74,16 @@ grppha infile=m2.pha outfile=m2.grp.pha chatter=0 comm=" group min ${group_min} 
 ######################################################################
 # hack header - grppha overwrites
 
+if [[ $CONVERT_TO_CTR -eq 1 ]]
+then
 echo "POISSERR=                    T / Poissonian errors applicable" > header.tmp
 for i in m1-${bgid}.grp.pha m2-${bgid}.grp.pha m1.grp.pha m2.grp.pha
 do
     fmodhead $i header.tmp
 done
 rm header.tmp
-
 # sleep 200
+fi
 
 ######################################################################
 # do the fitting
@@ -84,6 +92,7 @@ echo -e "
 data 1:1 m1.grp.pha
 data 2:2 m2.grp.pha
 
+cosmo 70 0 0.7
 query yes
 abund angr
 
@@ -124,8 +133,8 @@ pl ld res
 # weight churazov
 # weight model
 
-# statistic cstat
-statistic chi
+statistic cstat
+# statistic chi
 
 fit 100000000
 fit 100000000
